FlatUX Universal Theme
Per conversations on #341 and #391, I am submitting FlatUX, that can be used as the basis for a universal theme, compatible with XMB, Ozone, GLUI, etc.  Currently, this includes all source SVGs. The PNGs included at this time are intended to be replaced with files that will be created automatically by batch script.  FlatUX is intended to merge the FlatUI and RetroActive themes, once universal themes are officially rolled out.