Add FlatUX Theme
The files for FlatUX have been part of the repository for the last year, waiting for a universal theme format to be implemented and integrated.

Rather than wait for that work to be completed, I've decided to submit FlatUX as its own theme for end users to select.  Once this PR is merged, I will submit a PR on the main Retroarch repository to get the theme added to the list.

Long term, this theme is intended to replace both the flatui and the Retroactive themes.  For now, it will stand alongside flatui and the community can decide if it wishes to retire flatui in favor of FlatUX.

Once FlatUX is working and has been officially released, the Retroactive and Neoactive themes will officially be retired and their files will be removed from the repository in a future pull release.