(Directory) Renamed GamePark Handhelds
To be consistent with No-Intro naming, the files for the GP2X and Caanoo have been renamed.  The word "Holdings" has been removed.

Game Park created the GP32 but went bankrupt in 2007.  GamePark Holdings was founded by former employees who released the GP2X, GP2X Wiz and Caanoo.

Originally, the files were named to reflect the fact that two different companies released these handhelds.  No-Intro is not making that distinction at this time, so this PR will make it so Libretro theme icons are consistent with the DAT-O-Matic.